+++ internal/venice-common/src/test/java/com/linkedin/venice/pushmonitor/OfflinePushStatusTest.java
@@ -22,10 +22,8 @@ import static com.linkedin.venice.pushmonitor.ExecutionStatus.TOPIC_SWITCH_RECEI
 import static com.linkedin.venice.pushmonitor.ExecutionStatus.UNKNOWN;
 import static com.linkedin.venice.pushmonitor.ExecutionStatus.WARNING;
 import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertNull;
-import static org.testng.Assert.assertTrue;
 import static org.testng.Assert.fail;
 
 import com.linkedin.venice.exceptions.VeniceException;
@@ -214,30 +212,6 @@ public class OfflinePushStatusTest {
           }
           testInvalidTargetStatus(fromStatus, toStatus);
         }
-      }
-    }
-  }
-
-  @Test
-  public void testisDVCIngestionError() {
-    for (ExecutionStatus status: ExecutionStatus.values()) {
-      if (status == DVC_INGESTION_ERROR_DISK_FULL || status == DVC_INGESTION_ERROR_MEMORY_LIMIT_REACHED
-          || status == DVC_INGESTION_ERROR_OTHER) {
-        assertTrue(status.isDVCIngestionError(), status + " should not pass isDVCIngestionError()");
-      } else {
-        assertFalse(status.isDVCIngestionError(), status + " should pass isDVCIngestionError()");
-      }
-    }
-  }
-
-  @Test
-  public void testisIngestionError() {
-    for (ExecutionStatus status: ExecutionStatus.values()) {
-      if (status == ERROR || status == DVC_INGESTION_ERROR_DISK_FULL
-          || status == DVC_INGESTION_ERROR_MEMORY_LIMIT_REACHED || status == DVC_INGESTION_ERROR_OTHER) {
-        assertTrue(status.isIngestionError(), status + " should not pass isIngestionError()");
-      } else {
-        assertFalse(status.isIngestionError(), status + " should pass isIngestionError()");
       }
     }
   }
